c262 parity: fix array addition CI checks
